third_party/blink/web_tests/platform/mac-mac15-arm64/external/wpt/webxr/hit-test/ar_hittest_subscription_states_regular.https-expected.txt

This is a testharness.js-based test.
[FAIL] Hit test subscription succeeds if the feature was requested - webgl
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
[FAIL] Hit test subscription succeeds if the feature was requested - webgl2
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
[FAIL] Hit test subscription fails if the feature was not requested - webgl
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
[FAIL] Hit test subscription fails if the feature was not requested - webgl2
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
[FAIL] Hit test subscription fails if the feature was requested but the session already ended - webgl
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
[FAIL] Hit test subscription fails if the feature was requested but the session already ended - webgl2
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
Harness: the test ran to completion.
